etl/projects/learn-pyspark/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
psycopg2-binary==2.9.10
    # via -r requirements.in
py4j==0.10.9.7
    # via pyspark
pyspark==3.5.3
    # via -r requirements.in
